Add simple baseline run graphing
Add visual representation of baselines in graphs.  When passing a "baseline.<n>=<run_num>" argument to graph, a horizontal baseline will be drawn on the graph derived from the run passed.  The same test and metric are used for each plot.  The baseline is drawn as a thick darker line, in the same color as the original plot. The intended use is to compare something like trunk to the previous release.  This setting is off by default.